pyproject.toml
[project]
name = "folitools"
version = "0.3.1"
description = "Foli-seq data processing and analysis tools"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.85",
    "cyclopts==3.22.5",
    "joblib>=1.5.1",
    "matplotlib>=3.10.3",
    "matplotlib-venn>=1.1.2",
    "natsort>=8.4.0",
    "networkx>=3.5",
    "numpy>=2.3.1",
    "pandas>=2.3.0",
    "polars>=1.31.0",
    "pyarrow>=21.0.0",
    "pysam>=0.23.3",
    "scikit-learn>=1.7.0",
    "scipy>=1.16.0",
    "seaborn>=0.13.2",
    "sgad>=1.0.1",
    "tqdm>=4.67.1",
    "umi-tools>=1.1.6",
    "wcmatch>=10.1",
    "openpyxl",
    "primer3-py>=2.2.0",
    "pydeseq2>=0.5.3",
]

# --- uv dependency groups ---
[dependency-groups]
test = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
]
notebook = [
    "ipykernel>=6.29.5",
    "ipython>=9.3.0",
    "ipywidgets>=8.1.7",
]

# Make tests available in local runs by default (useful for contributors)
[tool.uv]
default-groups = ["test"]
package = true

# --- keep pip extras for non-uv users ---
[project.optional-dependencies]
test = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
]
notebook = [
    "ipykernel>=6.29.5",
    "ipython>=9.3.0",
    "ipywidgets>=8.1.7",
]

[project.scripts]
foli = "folitools.cli:app"
foli-primer = "folitools.primer_selection.cli:app"

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
# keep local data out of built dists
wheel-exclude  = [
    "src/folitools/primer_selection/data/**",
    "c/**",
    "rust/**",
]
source-exclude = [
    "src/folitools/primer_selection/data/**",
    "c/**",
    "rust/**",
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-v --tb=short --strict-markers -ra"
testpaths = ["tests"]
# keep as defaults: python_files, python_classes
# python_functions = ["test*"]      # allow testCamelCase
python_functions = ["test_*"]       # enforce underscore
xfail_strict = true

# Custom markers
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["src/folitools"]
branch = true
parallel = true
relative_files = true

[tool.coverage.report]
show_missing = true
skip_covered = true
# fail_under = 85  # For enforcing a bar in CI 
exclude_lines = [
  "pragma: no cover",
  "^def __repr__",
  "^if self\\.debug:",
  "^if settings\\.DEBUG",
  "^raise AssertionError",
  "^raise NotImplementedError",
  "^if 0:",
  "^if __name__ == ['\\\"]__main__['\\\"]:$",
  "^\\s*class .*\\(Protocol\\):",
  "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "htmlcov"
